+++ jsvg/src/main/java/com/github/weisj/jsvg/parser/impl/AttributeParser.java
@@ -460,25 +460,25 @@ public final class AttributeParser {
     /** {@code <number>}/{@code <length>}/{@code <percentage>} from one token, or {@code null} if it is neither. */
     private @Nullable Length lengthFromToken(@Nullable ComponentValue token, @NotNull PercentageDimension dimension) {
         if (token instanceof Token.Number) {
-            return Unit.RAW.valueOf(((Token.Number) token).value());
+            return Unit.RAW.valueOf((float) ((Token.Number) token).value());
         }
         if (token instanceof Token.Dimension) {
             Token.Dimension dimensionToken = (Token.Dimension) token;
             Unit unit = Unit.fromNonPercentageSuffix(dimensionToken.unit().toLowerCase(Locale.ENGLISH));
             if (unit == null) return null;
-            return unit.valueOf(dimensionToken.value());
+            return unit.valueOf((float) dimensionToken.value());
         }
         if (token instanceof Token.Percentage) {
             Unit unit = dimension.unit();
             if (unit == null) return null;
-            return unit.valueOf(((Token.Percentage) token).value());
+            return unit.valueOf((float) ((Token.Percentage) token).value());
         }
         return null;
     }
 
     public float parseFloat(@NotNull List<@NotNull ComponentValue> tokens, float fallback) {
         ComponentValue token = singleToken(tokens);
-        if (token instanceof Token.Number) return ((Token.Number) token).value();
+        if (token instanceof Token.Number) return (float) ((Token.Number) token).value();
         return fallback;
     }
 
@@ -490,14 +490,14 @@ public final class AttributeParser {
     /** {@code <angle>} from one token; bare {@code <number>} = degrees. */
     private static @Nullable Angle angleFromToken(@Nullable ComponentValue token) {
         if (token instanceof Token.Number) {
-            return new Angle(AngleUnit.Raw, ((Token.Number) token).value());
+            return new Angle(AngleUnit.Raw, (float) ((Token.Number) token).value());
         }
         if (token instanceof Token.Dimension) {
             Token.Dimension dimension = (Token.Dimension) token;
             String suffix = dimension.unit().toLowerCase(Locale.ENGLISH);
             for (AngleUnit unit : AngleUnit.units()) {
                 if (unit != AngleUnit.Raw && unit.suffix().equals(suffix)) {
-                    return new Angle(unit, dimension.value());
+                    return new Angle(unit, (float) dimension.value());
                 }
             }
         }
@@ -637,18 +637,18 @@ public final class AttributeParser {
     private @Nullable Length number(@NotNull List<@NotNull ComponentValue> args, int index) {
         if (index >= args.size()) return null;
         ComponentValue token = args.get(index);
-        return token instanceof Token.Number ? Unit.RAW.valueOf(((Token.Number) token).value()) : null;
+        return token instanceof Token.Number ? Unit.RAW.valueOf((float) ((Token.Number) token).value()) : null;
     }
 
     /** Transform angle argument in degrees: {@code <number>} or {@code <angle>}; null if neither. */
     private @Nullable Length angle(@NotNull List<@NotNull ComponentValue> args, int index) {
         if (index >= args.size()) return null;
         ComponentValue token = args.get(index);
-        if (token instanceof Token.Number) return Unit.RAW.valueOf(((Token.Number) token).value());
+        if (token instanceof Token.Number) return Unit.RAW.valueOf((float) ((Token.Number) token).value());
         if (token instanceof Token.Dimension) {
             Token.Dimension dimension = (Token.Dimension) token;
             AngleUnit unit = AngleUnit.fromSuffix(dimension.unit());
-            return unit != null ? Unit.RAW.valueOf(unit.toDegrees(dimension.value())) : null;
+            return unit != null ? Unit.RAW.valueOf(unit.toDegrees((float) dimension.value())) : null;
         }
         return null;
     }
@@ -668,10 +668,10 @@ public final class AttributeParser {
             @Nullable Percentage fallback, float min, float max) {
         ComponentValue token = singleToken(tokens);
         if (token instanceof Token.Percentage) {
-            return new Percentage(clamp(min, max, ((Token.Percentage) token).value() / 100f));
+            return new Percentage((float) clamp(min, max, ((Token.Percentage) token).value() / 100.0));
         }
         if (token instanceof Token.Number) {
-            return new Percentage(clamp(min, max, ((Token.Number) token).value()));
+            return new Percentage((float) clamp(min, max, ((Token.Number) token).value()));
         }
         return fallback;
     }
@@ -709,7 +709,7 @@ public final class AttributeParser {
         return token instanceof Token.Ident ? ((Token.Ident) token).name() : null;
     }
 
-    private static float clamp(float min, float max, float value) {
+    private static double clamp(double min, double max, double value) {
         return Math.max(min, Math.min(max, value));
     }
